Move helpers to src/helpers/
