+++ src/helpers/noop.js
